Reverted "Added unit test helper to write Agent log to Xunit output"
