--- src/Elastic.Apm/Logging/IAsyncLineWriter.cs
@@ -1,9 +1,0 @@
-using System.Threading.Tasks;
-
-namespace Elastic.Apm.Logging
-{
-	internal interface IAsyncLineWriter
-	{
-		Task WriteLineAsync(string line);
-	}
-}
