+++ test/Elastic.Apm.Tests/Elastic.Apm.Tests.csproj
@@ -25,7 +25,6 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="HelpersTests" />
-    <Folder Include="Helpers" />
     <Folder Include="Mocks\" />
     <Folder Include="ApiTests\" />
   </ItemGroup>
